FileTransferScriptLogs
At my job, we had an issue where a production machine was not communicating with server "X" due to problems with the operating system. However, it was communicating with the other machines connected to the same switch. So, as a workaround, I created this script. I found a machine on the switch that was communicating with the server and had the script collect the necessary data from the machine with the issue and transfer it directly to the server using a "L" communication. Additionally, the script displays the information "time, date, and quantity of files sent" on the terminal and saves this information in a log folder.